+++ GaudiKernel/src/Lib/ModuleInfo.cpp
@@ -1,5 +1,5 @@
 /***********************************************************************************\
-* (c) Copyright 1998-2019 CERN for the benefit of the LHCb and ATLAS collaborations *
+* (c) Copyright 1998-2021 CERN for the benefit of the LHCb and ATLAS collaborations *
 *                                                                                   *
 * This software is distributed under the terms of the Apache version 2 licence,     *
 * copied verbatim in the file "LICENSE".                                            *
@@ -114,7 +114,7 @@ System::ModuleType System::moduleType() {
   static ModuleType type = UNKNOWN;
   if ( type == UNKNOWN ) {
     const std::string& module = moduleNameFull();
-    int                loc    = module.rfind( '.' ) + 1;
+    size_t             loc    = module.rfind( '.' ) + 1;
     if ( loc == 0 )
       type = EXECUTABLE;
     else if ( module[loc] == 'e' || module[loc] == 'E' )
